files modified

+++ model_graph_cap.py
@@ -10,6 +10,7 @@ import tensorflow as tf
 
 from zoo_capsules import capsule_layer
 
+
 class ModelGraph():
     
     @staticmethod
@@ -39,7 +40,8 @@ class ModelGraph():
             emb_mat = tf.get_variable('embedding',
                                       [settings.vocab.size(), settings.vocab.emb_dim],
                                       initializer=tf.constant_initializer(settings.vocab.embeddings),
-                                      trainable = settings.emb_tune)
+                                      trainable = settings.emb_tune,
+                                      dtype=tf.float32)
             seq_emb = tf.nn.embedding_lookup(emb_mat, input_x)
             
             seq_mask = tf.cast(tf.cast(input_x, dtype = tf.bool), dtype = tf.int32)
